Update convert_to_onnx.ipynb
tf2onnx cannot convert the end2end model as it fails to convert the text vectorizer in keras. Hence, as huggingface, we can only convert the main inference model (glove-cnn) to onnx, and rely the keras text vectorizer to convert string input to int index for embedding layer at the main model.

Finish typing the script.