WIP EDA for topic modeling dataset
Update the duplicate selection. Further reduce the number of comments. (That is discovered at manually looking at the topi representation docs in LDA,, and found that some docs have the same comment and name of the game). Remove duplicate based on app_id, the text, the score, and the votes

EDA on frequent words in game reviews only. As shown in previous studies, a custom stopword list may be created particularly for game reviews (including words like 'game', 'to', 'go', 'get', 'got')