Deploy bertopic demo model for experimenting
A 40 topics bertopic model is deployed. Sending a review/comment to the model will return a topic id for that model (not that the id is only meaningful to the model generated that, as the relationship btw a numerical id and the topic is not consistent btw diff topic models).

Slight version difference in scikit-learn and scipy in the conda env file compared to that in development. scikit-learn: 1.3.0 -> 1.4.0; scipy 1.11.4 -> 1.12.0. No compatibility issue is found during local docker testing.