Add support to split sentences by token length
There are reviews that its token length may exceed the token length limit of the sbert model, resulting in tokens being truncated.

To handle that, reviews are split according to the token limits. Tokens without truncation are first generated, then spitted by the limit of the sbert model. Then the split reviews are regenerated from the tokens. Although the spacing btw words are not preserved, the resulting sbert embedding are still the same (verified by np.assert_allclose())