feat: per-game TLDR generation
A function to generate per-game TLDR generation. It currently accepts the name of the game as input (may be changed).

Switch from using docker chromadb client to use persistent client created from the script as I don't want to create a volume for the demo.

Add function description and moved some functions & variables to diff scripts for better readability in the main script for llm